Refactor flowtron in it's own module
