+++ npc_engine/services/tts/flowtron/text/__init__.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 """ from https://github.com/keithito/tacotron """
 import re
